requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
black==23.1.0
cachetools==5.3.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==39.0.1
Django==4.1.7
django-celery-beat==2.4.0
django-cors-headers==3.13.0
django-filter==22.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flower==1.2.0
grpcio==1.51.1
gunicorn==20.1.0
humanize==4.6.0
kombu==5.2.4
Markdown==3.4.1
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psycopg2-binary==2.9.5
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.1
six==1.16.0
sqlparse==0.4.3
tinkoff==0.1.1
tinkoff-investments==0.2.0b48
tornado==6.2
typing_extensions==4.5.0
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.6
